requirements.txt
streamlit
langchain
langchain-huggingface
python-dotenv
huggingface_hub